refactor full component and remove usless requests
